Qt: separate Action handling from Window class
